apsta_mode.sh: Fix invalid link assignment in RE-AP
Currently, during the bring-up of 2 GHz + 5 GHz Low + 6 GHz fronthaul AP
and 5 GHz High + 6 GHz backhaul STA, the fronthaul AP is moving to 5 GHz
Low + 5 GHz High + 6 GHz, which is incorrect.

This is happening, because if the link info is not found, it's is
assigning a junk link id.

In this change, added a check that, link id assignment will only take
place if and only if we get valid link info and add new api to get the
link id based on the control interface files of hostapd

Change-Id: Ibc12ef6e76257adfb13f1ce5991e788396e6f0df
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>
Signed-off-by: Rahul Bhattacharjee <quic_rbhattac@quicinc.com>