wifi: ath12k: Skip caldata download with module param
Currently checksum fails when caldata is corrupted and
that would lead to firmware assert. Add module params which
allows driver to skip the caldata download to recover the
eeprom with faulty calibration data. It is useful for initial
calibration. Interfaces cannot be brought up since the device
should be in FTM mode.

insmod ath12k.ko ftm_mode=1 skip_caldata=1

Change-Id: Ib293b07eed79f83c6af1780308fbb35da39abf6f
Signed-off-by: Aishwarya R <quic_aisr@quicinc.com>