wifi/mac80211.sh: Fix wifi config command
When wifi config command is given multiple times, it is stuck in
infinite loop.

Fix this by correcting the logic in while loop.

6G should always be in secured mode. Add encryption if band is
6G.

In json board file there is not variable name wlan, due to
this below warning is seen:
WARNING: Variable 'wlan' does not exist or is not an array/object
To resolve this comment part of code. There is no functionality
issue seen.

Change-Id: I01d8b2c199982e8416399e5b7ea24e3999e1ba41
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>
Signed-off-by: Maharaja Kennadyrajan <quic_mkenna@quicinc.com>