ath12k: do not use deflink storage for scan arvifs
The deflink storage link is used for fallback cases when proper
link is not known during Mgmt Tx. Currently the scan can use a
link id of 15 for ML cases which is reserved and when a arvif
is created, it can use up the deflink space causing the deflink's
link id to be set to 15. This may not get reassigned based on the
ML vap bringup sequence and can lead to usage of this link id 15
when mac80211 requests driver to send data/mgmt on any ML links.

Since this link id is reserved, driver on selection of the link id
from deflink space, it drops the packets. Hence to avoid this scenario
prevent usage of deflink space for scan purposes.

Change-Id: Ic8a2b653ae01f062f81b9a3b8060c9c23f1b457f
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>