apsta_mode.sh: Keep disable order reference while enabling
With the latest hostapd, dynamic link disable support is
added, but currently when we are disabling, the order
of the links enabled should atleast be preserved for
the last disabled link as this will be acted as
first_link_bss.

For example, if we have 3 link AP and we are disabling
the link one by one in the order say, 0->1->2 then
the enable order should start from 2.
i.e it can be either 2->1->0 or 2->0->1 but the last
disabled link should be enabled first.

Make changes in apsta script to modify the enable
sequence temporarily until full ML reconfigure support
is enabled in hostapd.

Change-Id: I4d9175b23f752771e59702af8d1d20aabac43b29
Signed-off-by: Manish Dharanenthiran <quic_mdharane@quicinc.com>