wifi: ath12k: Add support to download rxgainlut.bin
Update the QMI sequence to download rxgainlut bin files, if rxgainlut
support is enabled in firmware, firmware will advertise in the
firmware capabalities based on that rxgainlut.bin is downloaded to
firmware over QMI.

Rxgainlut binaries are used as input parameters for front end modules
(FEM) configuration.

If it fails to download rxgainlut from board-2.bin, it will check for
rxgainlut.b<board_id> file in /lib/firmware/ directory, if it fails to
download, then it will check for rxgainlut.bin file.

Change-Id: I10b0806adb40c677e7f8e834ecc7857f422e632b
Signed-off-by: Saleemuddin Shaik <quic_saleemud@quicinc.com>