ath12k: VHT capabilities updation after iw set antenna
After iw set antenna command execution the VHT capabilities
are not getting updated properly.

This is happening iw set antenna commmad is executed after
wifi down radio state is not ON. If radio state is not ON
VHT capabilities are not updated.

Updating the VHT Capabilities before the condition check -
if radio state is not ON.

Change-Id: Ia5a2692744d16b327a62fc297326df7f50d5d754
Signed-off-by: Harish Rachakonda <quic_rachakon@quicinc.com>