wifi: ath12k: Fix adding a QoS control right after dot11 header
The EAPOL 2/4 packets are dropped in the mac80211 driver
due to incorrect placement of the QoS control in the QoS
data frame.

For QoS data, the length of the ieee80211 header includes
QoS control (2 bytes) and HT control (4 bytes). In order
to build the ieee80211 QoS header, the QoS control field
should be placed in the appropriate position in the header.
Currently, when there is an HT control field in the header,
the QoS control field is added to the end which belonging to
the ht control field causing corruption of the header,
which causing EAPOL 2/4 packets to be dropped in mac80211.

To fix, use ieee80211_get_qos_ctl() to get the QoS control
pointer instead of using hdr_len.

Change-Id: Ie049e06c99ed43e1374efc1d8f5ba1f55e4ea7bd
Signed-off-by: Ramasamy Kaliappan <quic_rkaliapp@quicinc.com>
Signed-off-by: Dhanavandhana Kannan <quic_dhanavan@quicinc.com>