wifi: ath12k: Update the QMI sequence to download FW CFG
Add the new QMI message for downloading the firmware.dat file to
Target if the file is present in firmware path.

If the file is present, then set the flag in Host Capability QMI
request to indicate the Target about the firmware.dat file to be
downloaded in next QMI message. The Target will take the AFC data
from the FW CFG QMI message.

If the file is not present in firmware path, the flag will be
reset in Host Capability and the Target will take the default
data instead of getting it from the FW CFG QMI message.

Change-Id: I9a0dba331979c33c5b907a557cad86b566f1bf6a
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>