wifi: ath12k: Add support for SW Progressive DFS
When RADAR is detected in the operating bandwidth (OBW), device is
required to vacate the affected channels. This is typically done by
either reducing the bandwidth or by completely vacating the OBW.
In order to avoid changing the primary channel whenever possible, and
yet continue the ability to expand the bandwidth back, introduce the
SW based progressive DFS recipe. This allows the device to continue
radar detection on the affected channels by setting the device
bandwidth (DBW) same as before while reducing the OBW to avoid the
affected channels. Tx will be restricted to OBW while DBW will be used
for Rx.Hence RADAR detection will be active on DBW.
Background scan is started on DBW after NOP expires and channel is
moved from usable to available state. The channel will continue to
maintain its state as available until a RADAR is detected.

Change-Id: Iac59af45386f6c72b071eaca9d1c32bf438013ea
Signed-off-by: Shivani Tambatkar <quic_stambatk@quicinc.com>
Signed-off-by: Aloka Dixit <quic_alokad@quicinc.com>