wifi: ath12k: Handle enabling telemetry_agent module in openwrt upgrade.
Due to openwrt upgrade, the "wlan-telemetry" is moved to
"qsdk/qca/feeds/wlan_utils/", so removing it from here.

Currently a copy of the header files "telemetry_agent_wifi_driver_if.h"
and telemetry_agent_app_if.h are maintained for successful build of
telemetry related files with ath12k. Since the maintenance of these
file comes with higher cost, removing the copy of the files and it
will be included via the "EXTRA_CFLAGS" for a successful build.

Change-Id: Ib9cf3eec5ae93a149878ef8e785e12e848078579
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>