wifi: ath12k: Add Performance boost configs and smp affinity for RDP479 and RDP481
Currently there are no performance commands supported in the boost
performance.sh script as well as smp affinity settings for RDP479 and RDP481.

Update the dp irq names for QCN6432 as there is a conflict in
irq names in case of RDP479 and RDP481.
In ath12k driver irq names for PCI is in the "pci%u_xxxx_%d" where
%u is based on bus_id for PCI. For QCN6432 %u is based on
userpd_id.
In case of RDP479 and RDP481 we have one QCN9274 and one QCN6432,
which is leading to conflict as both the bus_id and userpd_id
is 1. Changed the irq_name of QCN6432 from "pci%u_xxxx_%d"
to "pcic%u_xxxx_%d".

Change-Id: Ia61e9d91ef9891303e3045ff794b5e4ab23644e3
Signed-off-by: Narendhar Reddy <quic_nare@quicinc.com>