wifi: ath12k: Enable UMAC reset support for DS mode
Add support for UMAC DP reset of partner SOC when
one SOC crashes in MLO as a part of MODE 1 SSR for DS Mode.

In this process shared memory will be used by Host and
Firmware to communicate the sequence of events. Host will
clear the data structures such as tx/rx descriptors and
re-initialize the rings and reset hp/tp to 0.

In addition to host handling, inform the umac reset progress
to PPE through PPE handlers so that PPE will take necessary
action to disable interrupts.

Change-Id: Ibc5d7c005de08d8fb484552cfb82cd8bec3c0c90
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>
Signed-off-by: Balamurugan S <quic_bselvara@quicinc.com>