apsta_mode.sh: Bring down STA vap when radar is detected
When a WDS repeater operating in a 5GHz DFS channel detects radar, but
not the connected Root AP, the AP VAP is brought down since switch
channel is not possible as the STA VAP is still connected to the Root AP.
There is no trigger to bring down the STA VAP in this case. When bringing
down the AP VAP, driver receives a stop_ap.
In the path of handling stop_ap, if there are other active VAPs and are
operating in a DFS channel, a multi-vdev restart is triggered. Since there
was no channel switch, the STA VAP is still active and connected to Root
AP on the radar affected channel. A restart on the STA VAP with the same
channel would result in a DFS violation error response from the FW.

In order to prevent this, the STA VAP should be brought down when there
is a radar detected. Add a script to listen to the radar-detected
message from the supplicant and bring down the STA VAP when it is
observed.

Change-Id: I8400b46e22b705dab0915c84499399649feba4b7
Signed-off-by: Amith A <quic_amitajit@quicinc.com>