wifi: ath12k: add support for SSR for IPQ5332+QCN6432+QCN9274.
Currently when an assert is happened with recovery set,
rootpd is stopped once all the radios are stopped.
This has a problem if the number of radios is not equal
to the number of userpds.

Make changes to stop the rootpd once all the userpds are
stopped.

Remove ath12k_core_rproc_boot_recovery api as it handles
only one userpd and one rootpd scenario. Make
ath12k_core_rpd_rproc_boot_recovery api as generic to handle
'n' number of userpd scenario.

Change-Id: Ia5d142600e73b14447dfc3f517d2814863394aef
Signed-off-by: Narendhar Reddy <quic_nare@quicinc.com>