wifi: mac80211: Fix station dump fields in WDS mode
After WDS AP gets associated with WDS STA, DTIM and beacon interval
values in the station dump is seen as 0.

During the WDS Association, DTIM and beacon interval stats are being referred
from the VLAN interface instead of the AP's interface. Fix this issue by
referring to the AP's interface to get the DTIM and beacon interval values.

Change-Id: I268c1243075879db0725c564ec3436126e202bb3
Signed-off-by: Ganesh Kariganuru Mahabalesh <quic_gkarigan@quicinc.com>