apsta_mode.sh: fix getting link id as well as hostapd ap status
Currently, at a few places, get_link_id() is used to get the first link id in
the interface and then use that to fetch the link status and in its output get
the actual number of links. However, num of links in disabled case would be 0
hence no action is taken further which is wrong.

Also, while getting hostapd ap status only one of the link is used in case of
MLO where as it is better to use all the links.

This fixes the above two issues.

Change-Id: Ie9d1d64ffc455870a85ab745d316938162455799
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>