wifi: mac80211: Add lock when accessing critical region
Lock is missed to add in function ieee80211_cfg_get_channel
while accessing the critical data like monitor chandef, hw wiphy.
Due to this lock miss, system get crashed in this function.
Because if reading and writing happen at the same time, there
is a chance of crash or unexpected behavior.

Lock should be added while accessing the critical region.

Fixes: 098baf231feb4 ("add mlo monitor support")

Change-Id: If3aff23a27d5710ecfc35e0b6622019d943cd19c
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>