wifi: ath12k: WAR Skip the driver STA access during SSR mode1
Currently if STA sends disassoc while MLO AP link is
in recovery, hostapd send peer deletion to driver.
During this peer deletion process, NULL pointer dereferenced,
because peer information is deleted as part of SSR mode1.

Fix this issue by adding a check whether radio is in recovery
and if the peer is ML peer and returning.

Change-Id: I67579d6cd8bb20c741656586c2d4672d692cfecd
Signed-off-by: Harish Rachakonda <quic_rachakon@quicinc.com>