wifi: ath12k: Fix channel state if radar is in non-obw
- Existing function, cfg80211_radar_event_device(), does not update the
channel state and the NOP timer when subsequent radar is injected in
device bandwidth. As per the feature if a subsequent radar is
detected in non-operating channel during an existing NOP the channel
state should be updated and the NOP timer should be set to the existing
NOP timer. This functionality is missing in
cfg80211_radar_event_device(). To resolve the issue a new variable
'time_nop_entered' is added. When NOP was started for the
non operating segment of the device bandwidth (due to an earlier radar
event) this variable will be used to set timer for any subsequent radar
hit channels in non-operating bandwidth. Until the ongoing NOP expires,
set dfs_state_entered for all new radar events on the non-operating
channels to be same as the tracked time because as per DFS rule AP
need not monitor these channels at all during NOP. Once NOP expires,
for the channel a CAC period ensures that radar detection is active
in non-operating channels before all channels are set to
NL80211_DFS_AVAILABLE.
- FW requires host to set both band_center_freq1 and band_center_freq2
for 160 MHz and 320 MHz to calculate correct offset when radar is
detected. Currently only the center frequency of complete device BW is
passed as band_center_freq1. Modified the logic to send both whenever
required.
- A new common function 'ath12k_wmi_check_device_present' is added in
wmi to check if valid device parameters are present instead of
existing checks. A similar check is updated in mac_op_set_background()
by calling the existing function cfg80211_chandef_device_present().

Fixes: 5010670 (wifi: ath12k: Add support for SW Progressive DFS)

Change-Id: Ic8c6a1848a89b2c6f708443ef301b441644a6eab
Signed-off-by: Shivani Tambatkar <quic_stambatk@quicinc.com>