wifi: ath12k: Add monitor vdev on Multi Vdev Restart WMI command
During channel switch, currently on MVR WMI command monitor vdev is not
included since it is not part of assigned_vif list and separate vdev
stop and start WMI commands are sent for monitor vdev without stopping
all other VDEVs. This is setting two parallel home channels on target
and causing the issue.

This change is to add monitor vdev on MVR WMI command and handle
monitor vdev on MVR event processing.

Change-Id: I0e25fe1d223df8b8b460ff8d0de82a6c58c15304
Signed-off-by: Rathees Kumar R Chinannan <quic_rrchinan@quicinc.com>