wifi: ath12k: fix max num peer for peer create
max number of peer is set to 145, when we have a
max vdev setup i.e. 16 vdevs, the expected number
of peers would be 128.

Here in the issue case we are hitting Firmware assert,
when we try to create the 129th peer, as part of peer
create the check is not valid which allows the creation
of 129th peer.

Updating the code to ensure only 128 peers are allowed.

Change-Id: Ieb1bf50f2a3fdd38a306056c37f5e154349b3dec
Signed-off-by: Amit Pandit <quic_pamit@quicinc.com>