wifi: ath12k: Fix channel state if AP is disabled multiple times
Existing code does not update the channel state after
NOP expires when AP is enabled and disabled multiple
times. This occurs because in function
cfg80211_is_wiphy_non_oper_device_chan(), the link ID
is uninitialized and hence set to default value 0. Added a
missing "for" loop for link by calling existing macro
for_each_link.
Added a check to verify if chandef->band is same as the
passed channel->band.
Updated the previous width check to verify that DBW parameters
are valid in chandef.

Change-Id: Id19068c9b33c6f2a07ddfe063e65337b24ec90f1
Signed-off-by: Shivani Tambatkar <quic_stambatk@quicinc.com>