wifi:ath12k: fix 6GHz hidden ssid scan issue
Currently when use hidden ssid in 6 GHz, the station vap is not bringup
due to scan does not fatch the hidden ssid.

The broadcast probe request sent by FW during the scan could have
been closer to the end of dwell time (30msec). Channel switch time
in WKK is 30 msec and so the channel switch might will impact
the reception of the ucast probe response from AP.

The STA itself would have moved to the next channel and
hence did not ACK the ucast probe response.
So the FW root cause of the problem is low dwell time causing
this Ucast probe response not ACKed and hence suggestion to host
is to increase the dwell time to 70msec for all channels on the scan

Change-Id: Ie7556fd3e1d1309829851fe2424c1b75a7b98326
Signed-off-by: Ramasamy Kaliappan <quic_rkaliapp@quicinc.com>
(cherry picked from commit cb348d2fe44d1bb50f6542cbd64a181800c2a830)