wifi: ath12k: reset arvif address on unlink
Currently, when station does scan in default link in default address
and later when it changes the link address during authentication
in mac80211 based on link_id seen probe response, this new link
address is not being updated to arvif's address.

Because of this, the auth request is sent out by station has different
source address where as in firmware vdev was created with arvif's address.
This difference in created vdev address in firmware and link address of
mac80211 leads to authentication response frame getting dropped as there
is no entry of link address of station in firmware.

Hence, add changes to reset the arvif's address during unlink as well as
update arvif's address based on link address sent from mac80211 except for
scan link id.

Fixes: I795ec6e6fcb ("wifi: ath12k: update scan bssid based on the scan request")

Change-Id: I9b84f6443185a96df8e5d0d955fe303fcc594ff7
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>