wifi: ath12k: Check channel state in case of valid puncturing bitmap
Existing function cfg80211_get_chans_dfs_usable() checks the
channel state for all the channels including any punctured channels
when device parameters are present. The ru_punct_bitmap
option in cli command is used for operating bandwidth.
When channels states are checked for device bandwidth set to
320 MHz and operating bandwidth set to 160 MHz the
channel state is checked for all the channels in 320 MHz
including the last 4 disabled channel. This issue is also
seen if a valid puncturing bit map is used for any other
bandwidth because the function uses a default 0 value.
Due to this the channel is never moved from usable to
available state after NOP expires. The issue is resolved
by updating the puncturing bitmap in the function
which depends on the position of operating
center frequency in device center frequency.

Change-Id: I574fe1ba9b51be19773ab137c90f0da596d84ba9
Signed-off-by: Shivani Tambatkar <quic_stambatk@quicinc.com>