wifi: ath12k: Handle STA state machine in assign vif chanctx for STA mode
The STA state machine is not handled for STA mode in the assign vif
chanctx ops in the driver when "disable_vht=1" wpa_supplicant config
used.

Due to this config, the firmware crash happens because of the peer is
deleted before the peer assoc command in the unassign vif chanctx ops.

Hence, fix this issue by handling the STA state machine to handle the
state machine sequence from the beginning state to completion for STA
mode when disable_vht wpa_supplicant config is used.

Change-Id: Iddcbc1af1707b20fc8e29f25df2e0ea5e25c4c17
Signed-off-by: Maharaja Kennadyrajan <quic_mkenna@quicinc.com>