wifi: mac80211: Set eht operation tag's basic mcs nss value.
Currently the ext tag: eht oper's basic nss value is set to the
max for both RX and TX. This is an issue because if the station's
RX and TX nss are below the basic nss of RX and TX of the eht
oper in AP's beacon, the association will not happen in EHT
mode(as per draft doc at EOM).

Hence the basic nss is set to 1 for both TX and RX instead of their
max nss values, so that the association between station and AP can
now be done with the EHT mode.

w.r.t. the draft doc.: IEEE 802.11-23/0638r4 "The basic EHT-MCS and
NSS set is the set of <EHT-MCS, NSS> tuples that are supported by all
EHT STAs that are members of an EHT BSS"

Change-Id: Ic72ba9d7cce96a8e5f1c0780ad1447818b98fd33
Signed-off-by: Naveen S <quic_naves@quicinc.com>