wifi: ath12k: fix wds broadcast packet drop
In 4 address mode, broadcast packets are dropped in
driver. This is because, tx link was chosen as primary
in ath12k_mac_get_tx_link() but in dp_tx it chose
assoc link's tcl_metadata and it was unable to find the
peer.

Add code changes to select primary link in dp_tx to send
broadcast packet in 4 address mode.

Fixes: I810f7e29ba7 ("wifi: ath12k: allow primary link selection")

Change-Id: Ic25fde62035510a7d8e18eba9da7380e75910164
Signed-off-by: Harshitha <quic_hprem@quicinc.com>