wifi: ath12k: Fix the cfg_params value to handle until 128 bits (3 bytes) to avoid the overflow.
Validating the input type by calculating the bit mask
and assign the bit value to the respective cfg parameters in
the ath12k_write_htt_stats_reset function. By this way it avoids
the overflow on the cfg parameters and it rolls over to the next
variable once it exceeds 32 bits. So, the htt stats reset.
type has capable to handle the input type until 128 bits.

Change-Id: Ib527d3d22aef05c4c7ee21cadc19e99737737ad2
Signed-off-by: Arulanbu Balusamy <quic_abalusam@quicinc.com>
Signed-off-by: Maharaja Kennadyrajan <quic_mkenna@quicinc.com>