wifi: mac80211: fix freeing STA in other link during stop_ap
While 5 GHz radio is doing CAC, clients in other links are getting
freed along with all STAs in that interface, which disrupts the
connection.

Send link_id to __sta_info_flush() function which removes the STAs
associated to that particular link which goes down or CAC during
stop_ap() function.

Fixes: I3fea698b7 ("wifi: mac80211: add link id argument for sta_flush() function")

Change-Id: I554515b2bf40e095b7d9bd1bcf2b88a5ebe1598f
Signed-off-by: Karthikeyan Kathirvel <quic_kathirve@quicinc.com>