ath12k: Fail to set mcast to ucast if sdata is not running
During wifi down, mcast to ucast call is invoked and at the
same time, remove interface is also called. In some corner
cases, mcast to ucast call is invoked after the interface is
removed. Due to this, warn on triggered as sdata in driver is
not available.

During wifi down, sdata stops running. So setting mcast to
ucast is pointless. To fix this issue, fail to set mcast to
ucast if sdata is not running.

Change-Id: I16b0401302b65ea1f2c3b1ab83ea98ca4525249a
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>