wifi: ath12k: update punct bitmap for peer assoc when device BW is set
Puncturing bitmap must be calculated over the device bandwidth
parameters when station is associated. Existing code uses the
same puncturing bitmap even when operating channel center frequencies
are greater than the device center frequency. Shift the puncturing
bitmap depending on the location of the operating bandwidth
in the device bandwidth.

Change-Id: I966fd17a4efc18ba64c10f4be3ae34d594bfe02a
Signed-off-by: Shivani Tambatkar <quic_stambatk@quicinc.com>