wifi: mac80211: do not deliver frames to PPE in RFS case
Currently, with RFS enabled in WDS mode, wifi rx packets are
given to PPE via ppe_vp_tx_to_ppe. This is causing PPE to
exception these packets and sent to bridge and then being forwarded
to infra-VAP of WDS AP which is causing unnecessary cpu
utilization. Fix this by not delivering frames to PPE when RFS
is enabled and instead delivering to network stack directly.

Fixes: I2c7d0f666f6 ("wifi: ath12k: Enable PPE RFS support")
Change-Id: Ieb921dc85e970d59d6f297069a23917e413d5949
Signed-off-by: Pradeep Kumar Chitrapu <quic_pradeepc@quicinc.com>