wifi: ath12k: call flush_work worker before flush the radios
There is a chance of mgmt tx workqueue and op_flush to go
mutually exclusive as both are acquiring the ah->mutex_lock.
During that time, if op_flush acquired the mutex_lock first
then the wait for event timeout would happen due to the wmi
mgmt tx completion is not received because of wmi mgmt tx
workqueue is waiting for the mutex_lock to be released by
op_flush.
This leads to transmit the pending mgmt packet after flush
which is not the correct way.

Hence, flush the workqueue in the op_flush before flush all
the radios with flush_work API (wait for the work to finish
executing the last queueing instance) to avoid the above scenario.

Change-Id: Ief77561c87ee08e1299a60e878720689860f8e75
Signed-off-by: Maharaja Kennadyrajan <quic_mkenna@quicinc.com>