wifi: ath12k: fix rssi value reporting of non-associated client
With Real RSSI feature, rssi_offset got added to RSSI reported
for peer of non-associated client.

This is not required since this RSSI is directly reported from HW
and don't need the current radio rssi offset to be added.

Reverting this change to fetch proper RSSI values for non-associated
client.

Fixes: Ie105a90d9319ad ("ath12k: add support for real signal strength")

Change-Id: Ibbdee1e0f9e36296c758e7979c4558b793f133ed
Signed-off-by: Sathishkumar Muruganandam <quic_murugana@quicinc.com>