ath12k: Fix save return value of func ath12k_mac_inc_num_stations
Currently when we are trying to associate more than 128 stations crash is
observed. This is because we don't support more than 128 stations and we
are trying to add more.

In code we have missed to save return value of ath12k_mac_inc_num_stations
func and due to this we are not able to identify that we have exceeded the
limit of num of stations.

Call trace:
 ath12k_core_issue_bug_on+0x38/0x1d4 [ath12k]
 ath12k_core_halt+0x96c/0xe5c [ath12k]
 process_one_work+0x1c4/0x29c
 worker_thread+0x24c/0x2ec
 kthread+0x104/0x114
 ret_from_fork+0x10/0x20
Code: b94b6823 39404c42 6b03005f 5400004c (d4210000)
---[ end trace 0000000000000000 ]---

Fixes: 2559fc6cc394 ("ath12k: Upgrade backports to backport-6.5-rc3")

Change-Id: Ia0c9b398ffe8bad00a8f6f46736d3c1158e7733e
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>