wifi: ath12k: Fix for the invalid memory access
Crash is observed when trying to access the htt tlv data memory address
because it is not present in the kernel page table.

when move from one tlv to another tlv, the size (tlv_len * 4) is
incremented with base pointer as it is defined as a u32 pointer variable.

Due to this, data pointer is moved to the invalid memory location which
is not present in the kernel page table.

To fix this issue, u8 pointer varaible is defined to move from one tlv
to another tlv.

Fixes: Ifcc6166908b(Add support for SAWF stats)

Change-Id: Ic480a73560e408f32aeaf4788c569009ece5d062
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>