wifi: ath12k: Set need_ptk_4_way flag to false for FT roaming
A few drivers may need to know if STA is using FT auth to decide on
allowing set key.

In order to allow such drivers to allow set key, add an NL attribute
NL80211_STA_FLAG_FT_AUTH to get the sta flag that denotes if it uses
FT authentication and then set the need_ptk_4_way flag to false if
this flag is set.

Change-Id: Id11ccaf3e73e1225157274999429a06a8cf47665
Signed-off-by: Amith A <quic_amitajit@quicinc.com>