wifi: mac80211: ath12k: Fix for use-after-free in key memory
During channel switch test case,  when switching from a
non-dfs channel to dfs channel 5G link goes down and keys
are uninstalled.

In this scenario, arvif is deleted and keys are cached in
set_key function as arvif is NULL when trying to unsinstall
the key.

After removing the links, key memory is freed in tear down
link function.

After vdev create, use-after-free observed when trying to
flush the cached key memory because this key memory is already
freed in tear down link function.

Currently while stopping a link AP, all keys from the
interface were removed. However with MLO there is a requirement
to free only the link keys.

Add changes to remove keys which are associated with the link
AP which is going to be stopped.

(Cherry-picked from
https://patchwork.kernel.org/project/linux-wireless/patch/20240227042251.1511122-2-quic_adisi@quicinc.com/)

Change-Id: Ib5b6a7e5304a5a9b3d7d17b373a9e5ef3e055c1b
Signed-off-by: Rameshkumar Sundaram <quic_ramess@quicinc.com>
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>