wifi: ath12k: reconfigure HTT Metadata for ext MSDU
HTT Metadata is currently being added to the beginning of the packet
skb which is causing VLAN pings and multicast traffic to be dropped
in Firmware.

This is due to HTT Metadata being utilized with ext MSDU. When used
with ext MSDU, HTT Metadata should be added to the ext MSDU skb
instead.

This fix removes functionality to add HTT Metadata to the beginning of
packet skb and adds functionality to add HTT Metadata to the end of
the ext MSDU skb.

Signed-off-by: Muna Sinada <quic_msinada@quicinc.com>
Change-Id: I5f0ae1ed4276f73a760c80ea0d856658c08e6842