QSDK: wifi: mac80211: Fix ieee80211 set active links function
Currently, in ieee80211_set_active_links does call drv_change_sta_links
and then call link_use_channel. The problem is that drv_change_sta_links
will update the resources to the firmware post recovery before the
creation or assignment of chanctx for the targeted vif which resulted
in unexpected behavior.

The caller must ensure that link are assigned with an existing or new
chanctx before changing the links resources in the target.

Change list:
        - Fix Mode-0 SSR for WDS STA mode - by calling link_use channel
        function before drv_change_sta_links.

Change-Id: I82a4eeb8a524cbd1a3bcb83ed538fb0ac42f0208
Signed-off-by: Sidhanta Sahu <quic_sidhanta@quicinc.com>