mac80211: [WAR] Fix deadlock during bring down using wdev specific flag
When links are brought down sequentially as a part of
__cfg80211_stop_ap, the worker threads currently check on link specific
is_going_down bit before sleeping on the sdata_lock. But if a specific
link is being brought down and the worker thread associated with a different
link gets execution context, the is_going_down will not be set for that
link; thus it sleeps on the sdata_lock, which ends up in a deadlock if
the process thread holding the wdev_lock waits for this worker thread to
finish.
Below is the call trace for such a deadlock.

Call trace:
[ 1197.513270] task:kworker/u8:1    state:D stack:0     pid:42    ppid:2      flags:0x00000008
[ 1197.516834] Workqueue: phy00 ieee80211_color_collision_detection_work [mac80211]
[ 1197.524926] Call trace:
[ 1197.532548]  __switch_to+0xe4/0xfc
[ 1197.534710]  __schedule+0x42c/0x5a4
[ 1197.538182]  schedule+0x58/0xac
[ 1197.541593]  schedule_preempt_disabled+0x20/0x38
[ 1197.544693]  __mutex_lock.constprop.0+0x260/0x4f8
[ 1197.549564]  __mutex_lock_slowpath+0x10/0x18
[ 1197.554155]  mutex_lock+0x4c/0x58
[ 1197.558494]  ieee80211_color_collision_detection_work+0x70/0x1ac [mac80211]
[ 1197.561732]  process_one_work+0x1c4/0x29c
[ 1197.568477]  worker_thread+0x24c/0x2ec
[ 1197.572653]  kthread+0x104/0x114
[ 1197.576288]  ret_from_fork+0x10/0x20

This change adds a wdev specific flag when the NETDEV_GOING_DOWN event
occurs. The cfg80211_leave sets the bit as soon as it
acquires the wdev lock. This will ensure the
ieee80211_color_collision_detection_work thread checks for this common
bit and decides not to take on the sdata lock, which will avoid the
deadlock.

Change-Id: I7114425d942c97e69b2d9cf1ecae51b21531a069
Signed-off-by: Ramanathan Choodamani <quic_rchoodam@quicinc.com>